Reorganize the Cabal packages around the library and compiler.
The biggest change of this package is to get rid of the `proto-lens-protoc`
package:
- Move the executable called `proto-lens-protoc` into `proto-lens.cabal`.
- Make a new package `proto-lens-setup` containing `Data.ProtoLens.Setup`.

This is a simpler organization that puts intertwined code from the library and
compiler into the same package.

Additionally, we rename the reexported modules (e.g., Data.ByteString =>
Data.ProtoLens.Reexport.Data.ByteString) so that other packages don't
accidentally depend on them being present.